# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Purple Sparrow API
  version: 0.1.0
  description: >
    Agent-native backend-as-a-service. Data + a deny-by-default policy engine,
    auth, object storage, edge functions (WASM), realtime (SSE), an agent layer,
    and an embedded dashboard — in a single Go binary. Every non-2xx response is
    the agent error envelope (code / message / remediation / next_actions / doc_ref).
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://127.0.0.1:8787
    description: Local (solo tier) default
security:
  - bearerAuth: []
...
paths:
...
  /v1/tables:
    post:
      tags: [Tables]
      operationId: createTable
      summary: Create a user table (admin). id and created_at are added automatically.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, columns]
              properties:
                name: { type: string, pattern: "^[a-z][a-z0-9_]*$" }
                columns: { type: array, items: { $ref: "#/components/schemas/Column" } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Table" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
    get:
      tags: [Tables]
      operationId: listTables
      summary: List table names (admin)
      responses: { "200": { description: OK }, "403": { $ref: "#/components/responses/Error" } }
  /v1/tables/{table}:
    parameters: [{ $ref: "#/components/parameters/Table" }]
    get:
      tags: [Tables]
      operationId: describeTable
      summary: Describe a table — columns and policies (admin)
      responses: { "200": { description: OK }, "403": { $ref: "#/components/responses/Error" }, "404": { $ref: "#/components/responses/Error" } }
    delete:
      tags: [Tables]
      operationId: dropTable
      summary: Drop a table (admin)
      responses: { "204": { description: Deleted }, "403": { $ref: "#/components/responses/Error" }, "404": { $ref: "#/components/responses/Error" } }
  /v1/tables/{table}/records:
    parameters: [{ $ref: "#/components/parameters/Table" }]
    post:
      tags: [Records]
      operationId: insertRecords
      summary: Insert one record (object) or many (array). Policy WITH CHECK applies.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - { type: object, additionalProperties: true }
                - { type: array, items: { type: object, additionalProperties: true } }
      responses:
        "201": { description: Created }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
    get:
      tags: [Records]
      operationId: queryRecords
      summary: "Query records (row-filtered by policy). PostgREST-style filters: ?col=op.value"
      parameters:
        - { name: select, in: query, schema: { type: string }, description: "Comma-separated columns" }
        - { name: order, in: query, schema: { type: string }, description: "col or col.desc, comma-separated" }
        - { name: limit, in: query, schema: { type: integer } }
        - { name: offset, in: query, schema: { type: integer } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, properties: { records: { type: array, items: { type: object, additionalProperties: true } } } } } } }
        "403": { $ref: "#/components/responses/Error" }
  /v1/tables/{table}/records/{id}:
    parameters: [{ $ref: "#/components/parameters/Table" }, { $ref: "#/components/parameters/Id" }]
    get:
      tags: [Records]
      operationId: getRecord
      summary: Get one record by id (policy-filtered)
      responses: { "200": { description: OK }, "403": { $ref: "#/components/responses/Error" }, "404": { $ref: "#/components/responses/Error" } }
    patch:
      tags: [Records]
      operationId: updateRecord
      summary: Patch a record (USING targets the row, WITH CHECK validates the result)
      requestBody: { required: true, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
      responses: { "200": { description: OK }, "400": { $ref: "#/components/responses/Error" }, "403": { $ref: "#/components/responses/Error" }, "404": { $ref: "#/components/responses/Error" } }
    delete:
      tags: [Records]
      operationId: deleteRecord
      summary: Delete a record within the policy USING scope
      responses: { "204": { description: Deleted }, "403": { $ref: "#/components/responses/Error" }, "404": { $ref: "#/components/responses/Error" } }
  /v1/policies:
    post:
      tags: [Policies]
      operationId: createPolicy
      summary: Create an access policy (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PolicyInput" }
      responses: { "201": { description: Created }, "400": { $ref: "#/components/responses/Error" }, "403": { $ref: "#/components/responses/Error" } }
  /v1/tables/{table}/policies:
    parameters: [{ $ref: "#/components/parameters/Table" }]
    get:
      tags: [Policies]
      operationId: listPolicies
      summary: List a table's policies (admin)
      responses: { "200": { description: OK }, "403": { $ref: "#/components/responses/Error" } }
  /meta:
    get:
      tags: [Agent]
      operationId: getMeta
      summary: One-call introspection — tables, columns, policies, counts (admin)
      responses: { "200": { description: OK }, "403": { $ref: "#/components/responses/Error" } }
...
  /docs/{slug}:
    get:
      tags: [Agent]
      operationId: getDoc
      summary: Fetch a documentation page (markdown)
      security: []
      parameters: [{ name: slug, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK, content: { text/markdown: {} } }, "404": { $ref: "#/components/responses/Error" } }
...
  /v1/realtime:
    get:
      tags: [Realtime]
      operationId: subscribeRealtime
      summary: Subscribe to row-change events over Server-Sent Events (policy-filtered per subscriber)
      parameters:
        - { name: table, in: query, required: false, schema: { type: string }, description: "Narrow to one table" }
      responses:
        "200":
          description: An event stream
          content:
            text/event-stream:
              schema: { $ref: "#/components/schemas/Event" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "An API key (ps_sk_…) or an RS256 access token as `Authorization: Bearer <value>` (or `X-API-Key`). Omit to act as anon."
  parameters:
    Table: { name: table, in: path, required: true, schema: { type: string } }
    Id: { name: id, in: path, required: true, schema: { type: string } }
...
  responses:
    Error:
      description: Agent error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: policy_denied }
            message: { type: string }
            remediation: { type: string }
            next_actions: { type: array, items: { type: string } }
            doc_ref: { type: string }
            request_id: { type: string }
    Column:
      type: object
      required: [name, type]
      properties:
        name: { type: string }
        type: { type: string, enum: [text, integer, real, boolean, timestamp, uuid, json] }
        nullable: { type: boolean }
        unique: { type: boolean }
    Table:
      type: object
      properties:
        name: { type: string }
        columns: { type: array, items: { $ref: "#/components/schemas/Column" } }
        created_at: { type: string }
    PolicyInput:
      type: object
      required: [table, action, roles]
      properties:
        table: { type: string }
        action: { type: string, enum: [select, insert, update, delete] }
        roles: { type: array, items: { type: string, enum: [anon, authenticated, project_admin] } }
        using: { type: string, description: "SQL-like row predicate (select/update/delete)" }
        check: { type: string, description: "WITH CHECK predicate (insert/update)" }
...
    Event:
      type: object
      properties:
        type: { type: string, enum: [insert, update, delete] }
        table: { type: string }
        id: { type: string }
        row: { type: object, additionalProperties: true }
        at: { type: string }